+++ vm/devices/virtio/virtio/src/tests.rs
@@ -1200,11 +1200,11 @@ impl VirtioDeviceV2 for FailingTestDevice {
         self.traits.clone()
     }
 
-    fn read_registers_u32(&mut self, _offset: u64) -> u32 {
+    fn read_registers_u32(&mut self, _offset: u16) -> u32 {
         0
     }
 
-    fn write_registers_u32(&mut self, _offset: u64, _val: u32) {}
+    fn write_registers_u32(&mut self, _offset: u16, _val: u32) {}
 
     fn start_queue(
         &mut self,
@@ -1257,11 +1257,11 @@ impl VirtioDeviceV2 for TestDevice {
         self.traits.clone()
     }
 
-    fn read_registers_u32(&mut self, _offset: u64) -> u32 {
+    fn read_registers_u32(&mut self, _offset: u16) -> u32 {
         0
     }
 
-    fn write_registers_u32(&mut self, _offset: u64, _val: u32) {}
+    fn write_registers_u32(&mut self, _offset: u16, _val: u32) {}
 
     fn start_queue(
         &mut self,
