+++ vm/devices/virtio/virtio_blk/src/lib.rs
@@ -324,7 +324,7 @@ impl VirtioDeviceV2 for VirtioBlkDevice {
         }
     }
 
-    fn read_registers_u32(&mut self, offset: u64) -> u32 {
+    fn read_registers_u32(&mut self, offset: u16) -> u32 {
         // The transport reads the device config space as a sequence of u32s.
         // We serialize VirtioBlkConfig to bytes and return the requested
         // 4-byte window. Three cases:
@@ -346,7 +346,7 @@ impl VirtioDeviceV2 for VirtioBlkDevice {
         }
     }
 
-    fn write_registers_u32(&mut self, _offset: u64, _val: u32) {
+    fn write_registers_u32(&mut self, _offset: u16, _val: u32) {
         // Config space is read-only for virtio-blk.
     }
