+++ vm/devices/virtio/virtio_net/src/lib.rs
@@ -279,7 +279,7 @@ impl VirtioDeviceV2 for Device {
         }
     }
 
-    fn read_registers_u32(&mut self, offset: u64) -> u32 {
+    fn read_registers_u32(&mut self, offset: u16) -> u32 {
         match offset {
             0 => u32::from_le_bytes(self.registers.mac[..4].try_into().unwrap()),
             4 => {
@@ -298,7 +298,7 @@ impl VirtioDeviceV2 for Device {
         }
     }
 
-    fn write_registers_u32(&mut self, _offset: u64, _val: u32) {}
+    fn write_registers_u32(&mut self, _offset: u16, _val: u32) {}
 
     fn start_queue(
         &mut self,
