+++ vm/devices/virtio/virtio_p9/src/lib.rs
@@ -84,7 +84,7 @@ impl VirtioDeviceV2 for VirtioPlan9Device {
         }
     }
 
-    fn read_registers_u32(&mut self, offset: u64) -> u32 {
+    fn read_registers_u32(&mut self, offset: u16) -> u32 {
         assert!(self.tag.len().is_multiple_of(4));
         assert!(offset.is_multiple_of(4));
 
@@ -100,7 +100,7 @@ impl VirtioDeviceV2 for VirtioPlan9Device {
         }
     }
 
-    fn write_registers_u32(&mut self, offset: u64, val: u32) {
+    fn write_registers_u32(&mut self, offset: u16, val: u32) {
         tracing::warn!(offset, val, "[VIRTIO 9P] Unknown write",);
     }
