+++ vm/devices/virtio/virtio_rng/src/lib.rs
@@ -67,11 +67,11 @@ impl VirtioDeviceV2 for VirtioRngDevice {
         }
     }
 
-    fn read_registers_u32(&mut self, _offset: u64) -> u32 {
+    fn read_registers_u32(&mut self, _offset: u16) -> u32 {
         0
     }
 
-    fn write_registers_u32(&mut self, _offset: u64, _val: u32) {}
+    fn write_registers_u32(&mut self, _offset: u16, _val: u32) {}
 
     fn start_queue(
         &mut self,
